webkit/data/layout_tests/platform/chromium-mac/LayoutTests/fast/dom/xmlhttprequest-constructor-in-detached-document-expected.txt

CONSOLE MESSAGE: line 205: Uncaught TypeError: undefined is not a function
Text for bug 25290: Crash when constructing XMLHttpRequest in a detached document.

PASS

